+++ .pre-commit-config.yaml
@@ -274,13 +274,13 @@ repos:
   # terraform modules via :validate and :lint targets with hermetic providers.
   # See MODULE.bazel tf.download(mirror={...}) for provider configuration.
 
-  # TODO: Re-enable once enforce_bazel_tests runs via Bazel (language: system)
-  # instead of language: python, so it can import util.bazel.workspace.
+  # TODO: Finish or remove enforce_bazel_tests (experimental, now in x/enforce_bazel_tests/).
+  # Needs language: system to import shared modules. See x/enforce_bazel_tests/ENFORCE_BAZEL_TESTS.md.
   # - repo: local
   #   hooks:
   #     - id: enforce-bazel-tests
   #       name: enforce bazel test cache
-  #       entry: devinfra/precommit/enforce_bazel_tests.py
+  #       entry: x/enforce_bazel_tests/enforce_bazel_tests.py
   #       language: python
   #       additional_dependencies: [pygit2]
   #       pass_filenames: false
